Update WindowsHello.cpp
A proof of concept that using a USB hardware key without having to type out the full Windows Hello password or shortening the Windows Hello password and jeopardizing the local Windows PC login. An improvement to this proof of concept is to clear the session to force a full database login whenever the USB hardware key is removed from the USB port. Finally, as a feature improvement, to add the last few character feature of the database password with USB hardware key to quick unlock like Keepass2Android.